devide test into files
