--- test/fixtures/index.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-	<meta charset="UTF-8">
-	<title>GitHub · Build software better, together.</title>
-	<!-- <script src="https://code.jquery.com/jquery&#45;2.2.4.min.js"></script> -->
-</head>
-<body>
-    <h1 class="heading">Build software better, together.</h1>
-	
-	<nav>
-		<ul class="top-nav">
-			<li class="explore"><a href="https://github.com/explore">Explore</a></li>
-		</ul>
-
-		<form id="top_search_form">
-			<input type="text" id="js-command-bar-field" placeholder="Search or type a command">
-			<input type='submit'>
-		</form>
-	</nav>
-
-	<form class="js-form-signup-home">
-		<input type="text" class="textfield" name="user[login]" placeholder="Pick a username">
-		<input type="text" class="textfield" name="user[email]">
-		<input type="password" class="textfield" name="user[password]">
-	</form>
-
-	<input type="checkbox" id="test-checkbox" checked="checked">
-
-	<a href="terms-of-service.html">terms of service</a>
-</body>
-</html>
