--- test/fixtures/terms-of-service.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-	<meta charset="UTF-8">
-	<title></title>
-</head>
-<body>
-	Terms of service dummy page.
-
-	<div>
-		Consectetur nesciunt odio inventore nihil sit nemo? Ullam labore labore voluptatibus tempore eveniet? Corporis nesciunt laudantium est deserunt animi debitis autem voluptatum fugit ratione dolores sapiente reprehenderit. Nam iste blanditiis!
-	</div>
-</body>
-</html>
